+++ pom.xml
@@ -38,6 +38,12 @@
 			<version>1.9.0</version>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+          	<groupId>org.codehaus.groovy</groupId>
+          	<artifactId>groovy-all</artifactId>
+          	<version>1.7.0</version>
+        </dependency>
+		
 	</dependencies>
 
 	<build>
@@ -50,6 +56,40 @@
 					<target>1.6</target>
 				</configuration>
 			</plugin>
+			<plugin>
+			      <groupId>org.codehaus.gmaven</groupId>
+			      <artifactId>gmaven-plugin</artifactId>
+			      <version>1.2</version>
+			      <configuration> 
+			        <providerSelection>1.7</providerSelection> 
+			      </configuration> 
+			      <dependencies>
+			          <dependency>
+			            <groupId>org.codehaus.gmaven.runtime</groupId>
+			            <artifactId>gmaven-runtime-1.7</artifactId>
+			            <version>1.2</version>
+			            <exclusions>
+			              <exclusion>
+			                <groupId>org.codehaus.groovy</groupId>
+			                <artifactId>groovy-all</artifactId>
+			              </exclusion>
+			            </exclusions>
+			          </dependency>
+			          <dependency>
+			            <groupId>org.codehaus.groovy</groupId>
+			            <artifactId>groovy-all</artifactId>
+			            <version>1.7.0</version>
+			          </dependency>
+			      </dependencies>
+			      <executions>
+			          <execution>
+			              <goals>
+			                  <goal>compile</goal>
+			                  <goal>testCompile</goal>
+			              </goals>
+			          </execution>
+			      </executions>
+			  </plugin>
 		</plugins>
 	</build>
 </project>
