--- src/groovy/dsl/dsl/main/ReunionDsl.groovy
@@ -1,5 +1,0 @@
-package dsl.main
-
-class ReunionDsl {
-
-}
